break out resource from dao into separate
