+++ ExtendedLearning/src/main/java/com/skilldistillery/sdelp/data/ResourceDAO.java
@@ -1,0 +1,11 @@
+package com.skilldistillery.sdelp.data;
+
+import com.skilldistillery.sdelp.entities.Resource;
+
+public interface ResourceDAO {
+	
+	Resource addResource(Resource resource);
+	
+	Resource updateResource(Resource resource);
+
+}
