+++ ExtendedLearning/src/main/java/com/skilldistillery/sdelp/data/SdelDaoJpaImpl.java
@@ -26,20 +26,7 @@ public class SdelDaoJpaImpl implements SdelDao {
 
 	
 
-	@Override
-	public Resource addResource(Resource resource) {
-		em.persist(resource);
-		return resource;
-	}
 
-	@Override
-	public Resource updateResource(Resource resource) {
-		Resource managedResource = em.find(Resource.class, resource.getId());
-		managedResource.setTitle(resource.getTitle());
-		managedResource.setResourceUrl(resource.getResourceUrl());
-		managedResource.setActive(resource.getActive());
-		return managedResource;
-	}
 
 	@Override
 	public Content addContent(Content content) {
